feat: ref 훅 분리
